pyproject.toml
[tool.poetry]
name = "game-store"
version = "0.1.0"
description = ""
authors = ["Alexandre Klostermann <alexandre@aupi.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
prettyconf = "^2.2.1"
django-jazzmin = "^3.0.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
django-filter = "^24.2"
django-cors-headers = "^4.4.0"
django-username-email = "^2.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
